refactor: response and request packaging
