+++ src/main/kotlin/org/shangahi/sellio_backend/api/controller/StoreRatingController.kt
@@ -1,6 +1,6 @@
 package org.shangahi.sellio_backend.api.controller
 
-import org.shangahi.sellio_backend.api.dto.StoreRatingResponse
+import org.shangahi.sellio_backend.api.dto.response.StoreRatingResponse
 import org.shangahi.sellio_backend.service.StoreRatingService
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
