+++ src/main/kotlin/org/shangahi/sellio_backend/api/controller/UserInfoController.kt
@@ -1,6 +1,6 @@
 package org.shangahi.sellio_backend.api.controller
 
-import org.shangahi.sellio_backend.api.dto.UserInfoResponse
+import org.shangahi.sellio_backend.api.dto.response.UserInfoResponse
 import org.shangahi.sellio_backend.api.mapper.toResponse
 import org.shangahi.sellio_backend.service.UserService
 import org.springframework.beans.factory.annotation.Value
