+++ src/main/kotlin/org/shangahi/sellio_backend/api/dto/response/ProductCardResponse.kt
@@ -1,4 +1,4 @@
-package org.shangahi.sellio_backend.api.dto
+package org.shangahi.sellio_backend.api.dto.response
 
 import java.util.UUID
