+++ src/main/kotlin/org/shangahi/sellio_backend/api/dto/response/SubCategoryDTO.kt
@@ -1,4 +1,4 @@
-package org.shangahi.sellio_backend.api.dto
+package org.shangahi.sellio_backend.api.dto.response
 
 import java.time.Instant
 import java.util.*
