+++ src/main/kotlin/org/shangahi/sellio_backend/api/mapper/UserMapper.kt
@@ -1,9 +1,9 @@
 package org.shangahi.sellio_backend.api.mapper
 
-import org.shangahi.sellio_backend.api.dto.UserInfoResponse
+import org.shangahi.sellio_backend.api.dto.response.UserInfoResponse
 import org.shangahi.sellio_backend.entity.User
 
-fun User.toResponse(cdnEndpoint: String): UserInfoResponse {
+fun User.toResponse(): UserInfoResponse {
     return UserInfoResponse(
         id = id,
         firstName = firstName,
