+++ src/main/kotlin/org/shangahi/sellio_backend/service/FavoriteProductService.kt
@@ -1,6 +1,6 @@
 package org.shangahi.sellio_backend.service
 
-import org.shangahi.sellio_backend.api.dto.FavoriteProductsResponse
+import org.shangahi.sellio_backend.api.dto.response.FavoriteProductsResponse
 import org.shangahi.sellio_backend.api.mapper.toFavoriteProductsResponse
 import org.shangahi.sellio_backend.repository.FavoriteProductRepository
 import org.springframework.stereotype.Service
