+++ src/main/kotlin/org/shangahi/sellio_backend/service/StoreRatingService.kt
@@ -1,6 +1,6 @@
 package org.shangahi.sellio_backend.service
 
-import org.shangahi.sellio_backend.api.dto.StoreRatingResponse
+import org.shangahi.sellio_backend.api.dto.response.StoreRatingResponse
 import org.shangahi.sellio_backend.repository.StoreRatingRepository
 import org.springframework.stereotype.Service
 import java.util.UUID
